backend-services/cloud-run/ai-agent/requirements.txt
google-adk
google-generativeai
google-cloud-storage
python-dotenv
Pillow